% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simon2stg.R
\name{EN}
\alias{EN}
\title{Title Expected N}
\usage{
EN(r1, n1, p, n)
}
\arguments{
\item{r1}{stopping response bar at stage-1}

\item{n1}{sample size stage-1}

\item{p}{true response rate}

\item{n}{sample size stage-2}
}
\value{

}
\description{
Title Expected N
}
