% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simon2stg.R
\name{PET}
\alias{PET}
\title{Title robability to stop for futility at IA}
\usage{
PET(r1, n1, p)
}
\arguments{
\item{r1}{stopping response bar at stage-1}

\item{n1}{sample size stage-1}

\item{p}{true response rate}
}
\value{
PET, probability to stop at stage-1
}
\description{
Title robability to stop for futility at IA
}
\examples{

}
