% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simon2stg.R
\name{ProbRej}
\alias{ProbRej}
\title{Title Prob for stop futility at end of the trial}
\usage{
ProbRej(r1, n1, p, r, n)
}
\arguments{
\item{r1}{stopping response bar at stage-1}

\item{n1}{sample size stage-1}

\item{p}{true response rate}

\item{r}{stopping response bar at stage-2}

\item{n}{sample size stage-2}
}
\value{

}
\description{
Title Prob for stop futility at end of the trial
}
