% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simon2stg.R
\name{Simon2stg_analysis}
\alias{Simon2stg_analysis}
\title{Title Function to implement Simon's 2-stage}
\usage{
Simon2stg_analysis(n1.s2stg, n.s2stg, r1.s2stg, r.s2stg, r.ia, r)
}
\arguments{
\item{n1.s2stg}{sample size stage-1}

\item{n.s2stg}{sample size stage-2}

\item{r1.s2stg}{stopping response bar at stage-1}

\item{r.s2stg}{stopping response bar at stage-2}

\item{r.ia}{observed number of response at end of stage-1}

\item{r}{observed total number of response at end of stage-2}
}
\value{
whether to reject null hypothesis (0 for fail to reject, 1 for to reject) and the sample size used
}
\description{
Title Function to implement Simon's 2-stage
}
\examples{
Simon2stg_analysis(n1.s2stg=20,n.s2stg=40,r1.s2stg=1,r.s2stg=13,r.ia=3,r=6)
}
